Endpoints: Move default endpoint getters from networks
